change max received message size to int.max to avoid error initialising binding

+++ VedaConnect/Client.cs
@@ -17,7 +17,7 @@ namespace VedaConnect
         public Client(string url, string username, string password)
         {
             var encoding = new TextMessageEncodingBindingElement { MessageVersion = MessageVersion.Soap11WSAddressing10 };
-            var transport = new HttpsTransportBindingElement { AuthenticationScheme = AuthenticationSchemes.Basic, KeepAliveEnabled = true, MaxReceivedMessageSize = long.MaxValue };
+            var transport = new HttpsTransportBindingElement { AuthenticationScheme = AuthenticationSchemes.Basic, KeepAliveEnabled = true, MaxReceivedMessageSize = int.MaxValue };
             var security = SecurityBindingElement.CreateUserNameOverTransportBindingElement();
             security.IncludeTimestamp = false;
